pyproject.toml
[project]
name = "django-langchained"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere==5.15.0",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "langchain>=0.3.27",
    "langchain-cohere>=0.4.4",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
